reorg(agents): split soc_agents into core/agents
Move services/soc_agents.py into a core/agents package (#482):
  builtins.py - AgentProfile + the 13 BUILTIN_AGENTS records
  prompts.py  - BASE_PROMPT, memory-palace block, render_base_prompt
  manager.py  - SOCAgentLibrary, AgentManager, CUSTOM_AGENT_ID_PREFIX

Fold the component-category map and the task-routing keyword table
into per-agent record fields, and unify the built-in and custom
profile builders into one build_profile. Behavior is unchanged: all
13 built-ins render identically and task routing is preserved.

Repoint every caller; no compatibility shim is left behind. The
frontend drops its hardcoded AGENT_META mirror for useAgentMeta
sourced from GET /agents, so labels and colors can no longer drift
from the backend. Update CLAUDE.md references.
