+++ backend/services/custom_agent_service.py
@@ -6,7 +6,8 @@ from typing import Any, Dict, List, Optional
 
 from database.connection import get_db_manager
 from database.models import CustomAgent
-from services.soc_agents import CUSTOM_AGENT_ID_PREFIX, render_base_prompt
+from core.agents.manager import CUSTOM_AGENT_ID_PREFIX
+from core.agents.prompts import render_base_prompt
 
 logger = logging.getLogger(__name__)
