+++ services/custom_workflow_service.py
@@ -46,7 +46,7 @@ def _validate_agent_ids(phases: List[Dict[str, Any]]) -> None:
     # Deferred to keep this service import-cheap for callers that only
     # want a .get() and don't touch the AgentManager.
     try:
-        from services.soc_agents import AgentManager
+        from core.agents.manager import AgentManager
 
         known = set(AgentManager().agents.keys())
     except Exception as e:
