+++ tests/unit/test_soc_agents_models.py
@@ -10,11 +10,8 @@ import pytest
 REPO = Path(__file__).resolve().parent.parent.parent
 sys.path.insert(0, str(REPO))
 
-from services.soc_agents import (  # noqa: E402
-    AgentProfile,
-    SOCAgentLibrary,
-    _BUILTIN_COMPONENT_CATEGORY,
-)
+from core.agents.builtins import AgentProfile, BUILTIN_AGENTS  # noqa: E402
+from core.agents.manager import SOCAgentLibrary  # noqa: E402
 
 pytestmark = pytest.mark.unit
 
@@ -35,6 +32,7 @@ def test_agent_profile_has_new_fields_with_safe_defaults():
 
 
 def test_builtin_categories_cover_all_built_ins():
+    expected_categories = {r["id"]: r["component_category"] for r in BUILTIN_AGENTS}
     agents = SOCAgentLibrary.get_all_agents()
     for agent_id, agent in agents.items():
         assert agent.component_category in {
@@ -42,9 +40,7 @@ def test_builtin_categories_cover_all_built_ins():
             "investigation",
             "reporting",
         }, f"{agent_id} has an invalid category: {agent.component_category}"
-        assert agent.component_category == _BUILTIN_COMPONENT_CATEGORY.get(
-            agent_id, "investigation"
-        )
+        assert agent.component_category == expected_categories[agent_id]
 
 
 def test_triage_agent_is_categorized_as_triage():
@@ -69,7 +65,7 @@ def test_custom_agent_builder_reads_model_and_category():
         "model": "claude-haiku-4-5-20251001",
         "component_category": "triage",
     }
-    profile = SOCAgentLibrary._build_from_custom(row)
+    profile = SOCAgentLibrary.build_profile(row)
     assert profile.model == "claude-haiku-4-5-20251001"
     assert profile.component_category == "triage"
 
@@ -81,6 +77,6 @@ def test_custom_agent_builder_defaults_category_when_missing():
         "role": "tester",
         "recommended_tools": [],
     }
-    profile = SOCAgentLibrary._build_from_custom(row)
+    profile = SOCAgentLibrary.build_profile(row)
     assert profile.model is None
     assert profile.component_category == "investigation"
